+++ Code/Engine/Math/Vector2.cpp
@@ -241,10 +241,10 @@ bool Vector2::SetFromText(const char* text)
 {
 	std::string stringText = std::string(text);
 
-	int commaPosition = static_cast<int>(stringText.find(","));
+	size_t commaPosition = stringText.find(",");
 
 	// No comma present in text
-	if (commaPosition == static_cast<int>(std::string::npos))
+	if (commaPosition == std::string::npos)
 	{
 		return false;
 	}
